--- a/benchmarks/fetch/headers-length32.mjs
+++ b/benchmarks/fetch/headers-length32.mjs
@@ -19,7 +19,7 @@ const headers = new Headers(
     'Width',
     'Accept-CH',
     'Via',
-    'Refresh',
+    'Set-Cookie',
     'Server',
     'Sec-Fetch-Dest',
     'Sec-CH-UA-Model',