--- a/lib/web/fetch/headers.js
+++ b/lib/web/fetch/headers.js
@@ -536,7 +536,7 @@ class Headers {
     const cookies = this[kHeadersList].cookies
 
     // fast-path
-    if (cookies === null) {
+    if (cookies === null || cookies.length === 1) {
       // Note: The non-null assertion of value has already been done by `HeadersList#toSortedArray`
       return (this[kHeadersList][kHeadersSortedMap] = names)
     }