**benchmarks/fetch/headers-length32.mjs**
- main
```
benchmark              time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------------- -----------------------------
Headers@@iterator   5'731 ns/iter     (4'300 ns … 516 µs)  5'300 ns 11'600 ns    115 µs
```

- this patch
```
benchmark              time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------------- -----------------------------
Headers@@iterator   5'151 ns/iter     (4'100 ns … 689 µs)  4'800 ns 11'100 ns    109 µs
```